Big fix on the trainer + moved to progress bar instead of spm style logs
- There was a big issue in the training, confusion about Node.id vs
Node.node_id, the first one is the id in the vocabulary, the second one
is the id within the lattice by order of insertion.
- Another missing function was prune_sentencespieces (renamed
prune_pieces here for clarity) which drops the less important pieces
at each loop.
- Switched to better progress bar.
- Dependency added: derive_builder for sanity with the builder paradigm
  of the trainer.
- Added the unicde script shenanigans from spm. It was easy to add to
check conformity with spm.
- We still don't have 100% same results as spm, because spm removes very
rare unicode characters like \u{3000} (in unigram_wagahaiwa_nekodearu.txt).
  I decided not to implement this as this is part of the preprocessing,
we could add this to the train from file test in `tests/unigram.rs`.
- In the same line of thought the default for split_by_whitespace is
  set to false here, because the pretokenizer should be doing the job
in `tokenizer`. It's not the case in spm. If we set it to true, the
sentences will be split by whitespace after the
`make_seed_sentence_pieces` function.